Adjust salvia::ext file locations.
