+++ CMakeLists.txt
@@ -10,6 +10,7 @@ find_package(range-v3 CONFIG REQUIRED)
 find_package(fmt CONFIG REQUIRED)
 find_package(GTest CONFIG REQUIRED)
 find_package(freeimage CONFIG REQUIRED)
+find_package(Freetype REQUIRED)
 # find_package(magic_enum CONFIG REQUIRED)
 include (cmake/IncludeBoost.cmake)
 include (cmake/IncludeLLVM.cmake)
