+++ salvia/include/salvia/ext/resource/mesh/material.h
@@ -1,0 +1,33 @@
+#pragma once
+
+#include "mesh.h"
+
+#include "salvia/resource/texture.h"
+
+#include "eflib/platform/typedefs.h"
+
+#include <string>
+
+namespace salvia::ext::resource {
+
+class obj_material : public attached_data {
+public:
+  obj_material();
+
+  std::string name;
+
+  eflib::vec4 ambient;
+  eflib::vec4 diffuse;
+  eflib::vec4 specular;
+
+  int shininess;
+  float alpha;
+
+  bool is_specular;
+
+  std::string tex_name;
+  salvia::resource::texture_ptr tex;
+  ~obj_material() {}
+};
+
+} // namespace salvia::ext::resource
