+++ samples/AnisotropicFilter/AnisotropicFilter.cpp
@@ -12,10 +12,10 @@
 #include <salvia/common/colors.h>
 #include <salvia/resource/texture.h>
 
-#include <salvia/ext/swap_chain/swap_chain.h>
-#include <salvia/ext/resource/mesh/sa/mesh_io.h>
-#include <salvia/ext/resource/mesh/sa/mesh_io_obj.h>
+#include "salvia/ext/resource/mesh/mesh_io.h"
+#include "salvia/ext/resource/mesh/mesh_io_obj.h"
 #include <salvia/ext/resource/texture/tex_io.h>
+#include <salvia/ext/swap_chain/swap_chain.h>
 
 #include <eflib/platform/main.h>
