+++ samples/ColorizedTriangle/ColorizedTriangle.cpp
@@ -1,5 +1,5 @@
-#include <salvia/ext/resource/mesh/sa/mesh_io.h>
-#include <salvia/ext/resource/mesh/sa/mesh_impl.h>
+#include "salvia/ext/resource/mesh/mesh_impl.h"
+#include "salvia/ext/resource/mesh/mesh_io.h"
 #include <salvia/ext/swap_chain/swap_chain.h>
 
 #include <salviau/include/common/sample_app.h>
