+++ samples/Font/Font.cpp
@@ -8,8 +8,8 @@
 #include <salvia/resource/texture.h>
 
 #include "font/font.h"
-#include <salvia/ext/resource/mesh/sa/mesh_io.h>
-#include <salvia/ext/resource/mesh/sa/mesh_io_obj.h>
+#include "salvia/ext/resource/mesh/mesh_io.h"
+#include "salvia/ext/resource/mesh/mesh_io_obj.h"
 #include <salvia/ext/resource/texture/tex_io.h>
 #include <salvia/ext/swap_chain/swap_chain.h>
