+++ samples/ObjLoader/ObjLoader.cpp
@@ -7,10 +7,10 @@
 #include <salvia/common/colors.h>
 #include <salvia/resource/texture.h>
 
+#include "salvia/ext/resource/mesh/material.h"
+#include "salvia/ext/resource/mesh/mesh_io.h"
+#include "salvia/ext/resource/mesh/mesh_io_obj.h"
 #include <salvia/ext/swap_chain/swap_chain.h>
-#include <salvia/ext/resource/mesh/sa/material.h>
-#include <salvia/ext/resource/mesh/sa/mesh_io.h>
-#include <salvia/ext/resource/mesh/sa/mesh_io_obj.h>
 
 #include <salviau/include/common/sample_app.h>
 #include <salviau/include/common/path.h>
