+++ samples/Sponza/Sponza.cpp
@@ -1,8 +1,8 @@
 
+#include "salvia/ext/resource/mesh/material.h"
+#include "salvia/ext/resource/mesh/mesh_io.h"
+#include "salvia/ext/resource/mesh/mesh_io_obj.h"
 #include <salvia/ext/swap_chain/swap_chain.h>
-#include <salvia/ext/resource/mesh/sa/mesh_io.h>
-#include <salvia/ext/resource/mesh/sa/mesh_io_obj.h>
-#include <salvia/ext/resource/mesh/sa/material.h>
 
 #include <salvia/shader/shader.h>
 #include <salvia/shader/shader_regs.h>
