+++ samples/Triangles/Triangles.cpp
@@ -6,9 +6,9 @@
 #include <salviar/include/rasterizer.h>
 #include <salvia/common/colors.h>
 
+#include "salvia/ext/resource/mesh/mesh_io.h"
+#include "salvia/ext/resource/mesh/mesh_io_obj.h"
 #include <salvia/ext/swap_chain/swap_chain.h>
-#include <salvia/ext/resource/mesh/sa/mesh_io.h>
-#include <salvia/ext/resource/mesh/sa/mesh_io_obj.h>
 
 #include <salviau/include/common/sample_app.h>
 #include <eflib/platform/main.h>
