+++ samples/VertexTextureFetch/VertexTextureFetch.cpp
@@ -7,9 +7,9 @@
 #include <salvia/common/colors.h>
 #include <salvia/resource/texture.h>
 
-#include <salvia/ext/swap_chain/swap_chain.h>
-#include <salvia/ext/resource/mesh/sa/mesh_io.h>
+#include "salvia/ext/resource/mesh/mesh_io.h"
 #include <salvia/ext/resource/terrain/gen_terrain.h>
+#include <salvia/ext/swap_chain/swap_chain.h>
 
 #include <salviau/include/common/sample_app.h>
 #include <eflib/platform/main.h>
